backend/de-password-manager/requirements.txt
anyio                     4.3.0
appnope                   0.1.4
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 2.4.1
async-lru                 2.0.4
attrs                     23.2.0
Babel                     2.14.0
beautifulsoup4            4.12.3
black                     24.8.0
bleach                    6.1.0
boto3                     1.34.153
botocore                  1.34.153
certifi                   2024.2.2
cffi                      1.16.0
charset-normalizer        3.3.2
click                     8.1.7
comm                      0.2.2
coverage                  7.6.1
cryptography              43.0.0
debugpy                   1.8.1
decorator                 5.1.1
defusedxml                0.7.1
distlib                   0.3.8
exceptiongroup            1.2.2
executing                 2.0.1
fastjsonschema            2.19.1
filelock                  3.15.4
fqdn                      1.5.1
h11                       0.14.0
httpcore                  1.0.5
httpx                     0.27.0
idna                      3.7
importlib_metadata        8.2.0
iniconfig                 2.0.0
ipykernel                 6.29.4
ipython                   8.18.1
ipywidgets                8.1.3
isoduration               20.11.0
jedi                      0.19.1
Jinja2                    3.1.3
jmespath                  1.0.1
json5                     0.9.25
jsonpointer               2.4
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
jupyter                   1.0.0
jupyter_client            8.6.1
jupyter-console           6.6.3
jupyter_core              5.7.2
jupyter-events            0.10.0
jupyter-lsp               2.2.5
jupyter_server            2.14.0
jupyter_server_terminals  0.5.3
jupyterlab                4.1.8
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.1
jupyterlab_widgets        3.0.11
MarkupSafe                2.1.5
matplotlib-inline         0.1.7
mistune                   3.0.2
moto                      5.0.12
mypy-extensions           1.0.0
nbclient                  0.10.0
nbconvert                 7.16.3
nbformat                  5.10.4
nest-asyncio              1.6.0
notebook                  7.2.1
notebook_shim             0.2.4
numpy                     1.26.4
overrides                 7.7.0
packaging                 24.0
pandas                    2.2.2
pandocfilters             1.5.1
parso                     0.8.4
pathspec                  0.12.1
pbr                       6.0.0
pexpect                   4.9.0
pip                       24.1
platformdirs              4.2.1
pluggy                    1.5.0
prometheus_client         0.20.0
prompt-toolkit            3.0.43
psutil                    5.9.8
ptyprocess                0.7.0
pure-eval                 0.2.2
pycparser                 2.22
Pygments                  2.17.2
pytest                    8.3.2
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
python-json-logger        2.0.7
pytz                      2024.1
PyYAML                    6.0.1
pyzmq                     26.0.2
qtconsole                 5.5.2
QtPy                      2.4.1
referencing               0.35.0
requests                  2.31.0
responses                 0.25.3
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.18.0
s3transfer                0.10.2
Send2Trash                1.8.3
setuptools                70.1.0
six                       1.16.0
sniffio                   1.3.1
soupsieve                 2.5
stack-data                0.6.3
stevedore                 5.2.0
terminado                 0.18.1
tinycss2                  1.3.0
tomli                     2.0.1
tornado                   6.4
traitlets                 5.14.3
types-python-dateutil     2.9.0.20240316
typing_extensions         4.12.2
tzdata                    2024.1
uri-template              1.3.0
urllib3                   2.2.1
virtualenv                20.26.3
virtualenv-clone          0.5.7
virtualenvwrapper         6.1.0
wcwidth                   0.2.13
webcolors                 1.13
webencodings              0.5.1
websocket-client          1.8.0
Werkzeug                  3.0.3
wheel                     0.43.0
widgetsnbextension        4.0.11
xmltodict                 0.13.0
zipp                      3.19.2